requirements.txt
black==24.10.0
click==8.1.7
isort==5.13.2
google-generativeai==0.8.3
pypdf2==3.0.1
ratelimitqueue==0.2.2
requests==2.32.3
tqdm==4.66.5
